Sun object
This code defines a sun object that can be given a longitude and
latitude, and used to calculate sunset time, sunrise time, solar noon,
and sunlight duration for the present day.

It also contains code for calculating the present time on a Wadokei, a
disused timekeeping method that started use in Edo era Japan. This way
of timekeeping basically divides day and night each into six equal
hours, so the clock hand moves at a different speed during night and
day.

More information on wadokei are available on wikipedia:
http://en.wikipedia.org/wiki/Japanese_clock

If the entire code is run, it will display a wadokei time for the
location entered on line 175.