error fix - mean longitude of the sun
So it turns out that the “longitude” of the sun is different form a
parameter called the “mean longitude” of the sun. Including the mean
longitude and using it to calculate topocentric coordinates corrected
some major errors in the program with calculating local sidereal time.
Now all planetary positions should be correct to within 7 arc minutes
or so! The elevation of mercury seems to be consistently off by several
arc minutes so I will keep looking for errors in the code. Inner
planets are supposed to be accurate to within about 1 arc minute
according to the original author of these algorithms. I probably messed
something up elsewhere!