minor edits
Added a calcMercuryDate function, and updated the eccentric anomaly
loop to be more accurate. The original instructions called for a
difference of 1E-5 degrees, but previously had it set to use radians.
Comparing to NASA horizons Ephemerides, position of mercury only seems
to be off by about 10 arc seconds! This could just be delay from light
travel time. I may also want to look into adjusting for the non
sphericity of the earth in the future.