Updates
Added tests for Luna, edited luna code to make it more testable and
useable while still maintaining the some of the previous string output
methods. Tests for luna check that the calculated position of the moon
stays within .03 degrees of expected value obtained from NASA’s
HORIZON’s ephemeris. Also checks to make sure that the riseset method
does not become stuck.

For the planets app, I added tests to check values for 40 years in the
future, and for an additional location in Japan.